Updated Menu
Fixed some bugs in the multiplayer and singleplayer menues.
Added a bunch of comments (some are probably badly written and/or
incorrect).
Moved saves folder from "My Documents" to "My
Documents\MainDigger\Saves".
Added a method that creates the save foler (of course).
Made some of the buttons in singleplayer/multiplayer allign with the
bottom right corner (instead of letting them float in the middle
somewhere).
Added a singleplayer browser for saved games in the saves folder (a
modified version of the server browser).

And some other, minor, stuffs that I do not remember.