Improved game broswers
Improved both singleplayer and multiplayer game browsers.
Fixed multiple crash causing bugs.
"Connect" and "Play" are now grayed out until a server/save is selected
(this is to make it more visable when you can and can not press these
buttons).
Now the visual server/save selection should be correct (it used to be
incorrect if you picked a save/server and then scrolled).
Commented out classes that are in the works (MenuWidgets) since I'm
planning on remaking the widget system (probably not, but mby).
Other stuffs.